Create a StreamManager
the StreamManager solves an issue where Discord would get really
out-of-sync when it tried to pass the stream to the client, since the
stream was initialized, regardless of when the user actually got around
to requesting the stream. This fixes that issue by initializing the
stream right when the user makes the command to access it.